backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydub==0.25.1
faster-whisper>=1.0.0
torch>=2.4.0,<2.6.0
torchaudio>=2.4.0,<2.6.0
numpy>=1.24.3
pandas>=2.0.3
scipy>=1.11.0
librosa>=0.10.0
soundfile>=0.12.1
python-dotenv>=1.0.0
aiofiles>=23.2.1
pyannote.audio>=3.1.0
matplotlib>=3.7.0

